+++ src/components/Footer/index.jsx
@@ -1,34 +1,81 @@
 import React from 'react';
-// import { NavLink } from 'react-router-dom';
+import Proptypes from 'prop-types';
+import { Link } from 'react-router-dom';
+
+import { connect } from 'react-redux';
+import { loadSocialMedia } from '../../store/base/base';
 
 import DonateButton from '../Donate Button';
 
 import './styles.scss';
 
-export class Footer extends React.Component {
+const mapDispatchToProps = dispatch => {
+  return {
+    loadSocialMedia: () => {
+      dispatch(loadSocialMedia('2umAjY8tMEQuOIYUmii08Y'));
+    }
+  };
+};
+
+const mapStateToProps = state => {
+  return { base: state.base };
+};
+
+class Footer extends React.Component {
   constructor() {
     super();
+
+    this.state = {
+      pages: ['projects', 'blog', 'team', 'about'],
+      socials: []
+    };
+
+    this.componentDidMount = this.componentDidMount.bind(this);
   }
+
+  componentDidMount() {
+    const { loadSocialMedia } = this.props;
+
+    // Dispatch loadSocialMedia()
+    loadSocialMedia();
+  }
+
   render() {
+    const { pages } = this.state;
+    const { socials } = this.props.base;
+
+    // Format footer page links
+    const footerPageLinks = pages.map(page => {
+      return (
+        <li key={page}>
+          <Link to={'/' + page}>{page}</Link>
+        </li>
+      );
+    });
+
+    // Format footer social media links
+    if (socials.length != 0) {
+      var footerSocialLinks = Object.keys(socials).map(key => {
+        return (
+          <li key={key}>
+            <a href={socials[key]} target="_blank" rel="noopener noreferrer">
+              {key}
+            </a>
+          </li>
+        );
+      });
+    }
+
     return (
       <div className="footer">
         <footer className=" container">
           <div className="row align-items-top">
             <div className="footer-widget col-4 d-none d-md-flex">
-              <ul className="list-unstyled">
-                <li>Projects</li>
-                <li>Blog</li>
-                <li>About</li>
-                <li>Team</li>
-              </ul>
+              <ul className="list-unstyled">{footerPageLinks}</ul>
             </div>
 
             <div className="footer-widget col-4 d-none d-md-flex">
-              <ul className="list-unstyled">
-                <li>Facebook</li>
-                <li>Instagram</li>
-                <li>Mail</li>
-              </ul>
+              <ul className="list-unstyled">{footerSocialLinks}</ul>
             </div>
 
             <div className="col-12 col-md-4 text-left">
@@ -59,3 +106,13 @@ export class Footer extends React.Component {
     );
   }
 }
+
+Footer.Proptypes = {
+  loadSocialMedia: Proptypes.func,
+  base: Proptypes.obj
+};
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(Footer);