+++ src/store/base/actions.js
@@ -2,7 +2,8 @@
 
 // Import Redux base types
 import {
-    LOAD_HEADER_IMAGE_SUCCESS
+    LOAD_HEADER_IMAGE_SUCCESS,
+    LOAD_SOCIALS_SUCCESS
 } from './types';
 
 // Action that returns the successfully loaded header image
@@ -11,4 +12,12 @@ export function loadHeaderImageSuccess(headerImage) {
         type: LOAD_HEADER_IMAGE_SUCCESS,
         headerImage
     };
+}
+
+// Action that returns successfully retreived socials
+export function loadSocialsSuccess(socials) {
+    return {
+        type: LOAD_SOCIALS_SUCCESS,
+        socials
+    }
 }