+++ src/store/base/reducer.js
@@ -2,7 +2,8 @@
 
 import initialState from '../initialState';
 import {
-    LOAD_HEADER_IMAGE_SUCCESS
+    LOAD_HEADER_IMAGE_SUCCESS,
+    LOAD_SOCIALS_SUCCESS
 } from './types';
 
 // Base app reducer
@@ -15,6 +16,13 @@ export default function baseReducer(state = initialState.base, action) {
                 ...state,
                 headerImage: action.headerImage
             };
+
+        // Successful socials load type
+        case LOAD_SOCIALS_SUCCESS:
+            return {
+                ...state,
+                socials: action.socials
+            }
 
             // Default case
         default: