+++ src/store/base/types.js
@@ -2,3 +2,6 @@
 
 // Successfully loaded header image
 export const LOAD_HEADER_IMAGE_SUCCESS = 'LOAD_HEADER_IMAGE_SUCCESS';
+
+// Successfully loaded socials
+export const LOAD_SOCIALS_SUCCESS = 'LOAD_SOCIALS_SUCCESS';