+++ src/store/initialState.js
@@ -1,6 +1,7 @@
 // Initial Redux state
 export default {
     base: {
+        socials: [],
         headerImage: ''
     },
     blog: {